Add config files to enable scons as build system.

+++ clear_cmake.sh
@@ -1,4 +1,27 @@
-rm CMakeCache.txt 
+if [ -f CMakeCache.txt ]
+then
+	rm CMakeCache.txt
+fi
+
+if [ -f cmake_install.cmake ]
+then
+	rm cmake_install.cmake
+fi
+
+if [ -f Makefile ]
+then
+	rm Makefile
+fi
+
+if [ -f .sconsign.dblite ]
+then
+	rm .sconsign.dblite
+fi
+
+if [ -f numsim ]
+then
+	rm numsim
+fi
+
 rm -rf CMakeFiles/
-rm cmake_install.cmake 
-rm Makefile 
+rm -rf build/
